requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
kafka-python==2.0.2
pydantic==1.10.11
jinja2==3.1.2
python-dotenv==1.0.0
pybars3
